# backend/server-a/api/swagger.yaml
...
swagger: '2.0'
info:
  description:
    "This is the Swagger 2.0 API for Web Architectures course group project\
    \ work.\n"
  version: '1.1.0'
  title: 'Make me a sandwich'
  contact:
    email: 'mikko.nurminen@tuni.fi'
  license:
    name: 'Apache 2.0'
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost'
basePath: '/v1'
...
schemes:
  - 'http'
paths:
  /order:
    get:
      tags:
        - 'order'
      summary: 'Get a list of all orders. Empty array if no orders are found.'
      operationId: 'getOrders'
      produces:
        - 'application/json'
      parameters: []
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/ArrayOfOrders'
      x-swagger-router-controller: 'Order'
    post:
      tags:
        - 'order'
      summary: 'Add an order for an sandwich'
      operationId: 'addOrder'
      produces:
        - 'application/json'
      parameters:
        - in: 'body'
          name: 'order'
          description: 'place an order for a sandwich'
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: 'Order not created'
      x-swagger-router-controller: 'Order'
  /order/{orderId}:
    get:
      tags:
        - 'order'
      summary: 'Find an order by its ID'
      description: 'IDs must be positive integers'
      operationId: 'getOrderById'
      produces:
        - 'application/json'
      parameters:
        - name: 'orderId'
          in: 'path'
          description: 'ID of the order that needs to be fetched'
          required: true
          type: 'string'
          minimum: 1
      responses:
        '200':
          description: 'successful operation'
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Order not found'
      x-swagger-router-controller: 'Order'
    put:
      tags:
        - 'order'
      summary: Update an order for an sandwich
      operationId: updateOrder
      produces:
        - application/json
      parameters:
        - name: 'orderId'
          in: 'path'
          description: 'ID of the order that needs to be updated'
          required: true
          type: 'string'
          minimum: 1
        - in: body
          name: 'updatedOrder'
          description: update an order for a sandwich
          required: true
          schema:
            $ref: '#/definitions/Order'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        '404':
          description: Order not found
...
definitions:
  Order:
    type: 'object'
    properties:
      id:
        type: 'string'
      sandwichId:
        type: 'integer'
        format: 'int64'
      status:
        type: 'string'
        description: 'Order Status'
        enum:
          - 'ordered'
          - 'received'
          - 'inQueue'
          - 'ready'
          - 'failed'
    xml:
      name: 'Order'
    example:
      sandwichId: 6
      id: 0
      status: 'ordered'
  ArrayOfOrders:
    type: 'array'
    items:
      $ref: '#/definitions/Order'
...